feat(build): qt6 port + qml / qt3d removal
This is a port from Qt 5 to Qt 6.

Included:
- Use new Qt 6 VCPKG and Qt 6 docker image
- Updated many dependencies due to VCPKG update
- Removed QML, Qt3D

Also:
- Removed deprecated boost::asio::io_service
- Fixed rare race condition in worker_asio
- Fixed selector
- Fixed http post test
- Fixed video_controls GUI tests
- Fixed cardinal layout main title
- Sequencer widget reimplementation
- Removed CMake warnings and deprecated code

+++ modules/ui/qt/model/model_series_list.cpp
@@ -1,6 +1,6 @@
 /************************************************************************
  *
- * Copyright (C) 2009-2024 IRCAD France
+ * Copyright (C) 2009-2025 IRCAD France
  * Copyright (C) 2012-2020 IHU Strasbourg
  *
  * This file is part of Sight.
@@ -116,7 +116,7 @@ void model_series_list::starting()
 
     m_tree = new QTreeWidget();
 
-    m_tree->setColumnCount(m_headers.size());
+    m_tree->setColumnCount(static_cast<int>(m_headers.size()));
     m_tree->setHeaderLabels(m_headers);
 
     if(m_enable_hide_all)
